Rename some files to match name convention
