+++ simulator/infra/replacement/cache_replacement.h
